New content!
Updated the Events index page to better reflect the direction we want
it to go in.  Updated the header bar to first off be better html, and
secondly be easier to style.  Added a quick test of the what we want
the user page to show, teams wise.  And last but not least did a quick
tweak to the way the menu is in code and removed some testing code from
that file.  1.0 seems really close and yet really far.