Going live soon :|
The site will be going live very soon which is a scary thought and lots
needs to be done.

I got a lot done to be ready though.  I added a mailer with a welcoming
email.  Also been working on styling issues.  The site is looking
better but more to come.

Also added the letter_opener gem so it's easier to debug mail in
development mode.  To make letter_opener work I had to mess with Mail
initializer.